pyproject.toml
[project]
name = "duckpi-ic"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = "~3.9"

dependencies = [
  "zaber-motion (>=7.3.0,<8.0.0)",
  "pyyaml (>=6.0.2,<7.0.0)",
  "schema (>=0.7.7,<0.8.0)",
]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
